backend/requirements.txt
pyramid==2.0.2
pyramid_jinja2==2.10
pyramid_debugtoolbar==4.9
pyramid_tm==2.5
SQLAlchemy==1.4.46
transaction==3.0.1
zope.sqlalchemy==1.6
waitress==2.1.2
bcrypt==4.0.1
PyJWT==2.6.0
marshmallow==3.19.0
webtest==3.0.0
pytest==7.3.1
pytest-cov==4.1.0
cornice==6.0.1